Cancel movement for the duration of the double-click delay.
When attempting to double (or even single) click, an involuntary
movement could move the mouse far enough to initiate a drag, which
can be particularly annoying with something like firefox, where
clicking on a tab activate it can instead drag it and create a new
browser window.

The current behavior cancels movement for the delay amount, regardless
of the button states (because an up and then another down-up may be
coming if this is an intended double-click.)  As a result, any click
by the user is followed by this delay before any movement is registered.
This will take some getting used to, but perhaps it's an acceptable
trade-off to avoiding unintentional behavior due to tremor.

It might be worth tracking movement during this time (while still
cancelling movement) with a large threshold, to allow intentional
large movement.  This might need to only happen *after* an initial
button release, to avoid cancelling out any benefit from the above
motion cancellation.

FIXME: The hardcoded double-click wait time should instead follow
the user's GtkSetting for this, except that we need to run as root,
and trying to instantiate a GtkSettings would load the root user's
instead.